requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
anyio==4.9.0
async-timeout==5.0.1
asyncio==3.4.3
attrs==25.3.0
certifi==2025.4.26
exceptiongroup==1.2.2
frozenlist==1.6.0
h11==0.16.0
httpcore==1.0.9
httpx==0.25.2
idna==3.10
multidict==6.4.3
propcache==0.3.1
python-telegram-bot==20.7
sniffio==1.3.1
typing-extensions==4.13.2
yarl==1.20.0
